requirements.txt
pandas
numpy
scikit-learn==1.7.2
matplotlib
seaborn
openpyxl
fastapi
uvicorn[standard]
prefect
pydantic
pydantic-settings
pytest
pytest-asyncio
deepchecks
httpx
requests
python-dotenv
pyarrow
prefect==3.6.5
xgboost
streamlit>=1.40.0
scikit-learn==1.7.2
